+++ hw/components/sl_tg/sl_tg_device.cpp
@@ -45,7 +45,7 @@ void sl_tg_device::reset_input ()
     fseek (tg_file, 0, SEEK_SET);
 }
 
-void sl_tg_device::write (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr)
+void sl_tg_device::write (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr)
 {
     switch (ofs)
     {
@@ -54,21 +54,21 @@ void sl_tg_device::write (unsigned long ofs, unsigned char be, unsigned char *da
         default:
             printf ("Bad %s:%s ofs=0x%X, be=0x%X, data=0x%X-%X!\n", name (),
                     __FUNCTION__, (unsigned int) ofs, (unsigned int) be,
-                                   (unsigned int) *((unsigned long*)data + 0), (unsigned int) *((unsigned long*)data + 1));
+                                   (unsigned int) *((unsigned int *)data + 0), (unsigned int) *((unsigned int *)data + 1));
             //exit (1);
             break;
     }
     bErr = false;
 }
 
-void sl_tg_device::read (unsigned long ofs, unsigned char be, unsigned char *data, bool &bErr)
+void sl_tg_device::read (unsigned int ofs, unsigned char be, unsigned char *data, bool &bErr)
 {
     int             i;
 	int				offset_dd = 0;
 	int				mod = 0;
 
-    *((unsigned long *)data + 0) = 0;
-    *((unsigned long *)data + 1) = 0;
+    *((unsigned int *)data + 0) = 0;
+    *((unsigned int *)data + 1) = 0;
 
 	switch (be)
 	{
@@ -112,7 +112,7 @@ void sl_tg_device::read (unsigned long ofs, unsigned char be, unsigned char *dat
         case 0x80:
             if (!tg_bytes_left)
                 reset_input ();
-            *((unsigned long *)data + 0) = (tg_bytes_left + 3) / 4;
+            *((unsigned int *)data + 0) = (tg_bytes_left + 3) / 4;
             break;
         default:
             printf ("Bad %s:%s ofs=0x%X, be=0x%X!\n",  name (), __FUNCTION__, (unsigned int) ofs, (unsigned int) be);
@@ -121,7 +121,7 @@ void sl_tg_device::read (unsigned long ofs, unsigned char be, unsigned char *dat
     bErr = false;
 }
 
-void sl_tg_device::rcv_rqst (unsigned long ofs, unsigned char be,
+void sl_tg_device::rcv_rqst (unsigned int ofs, unsigned char be,
                           unsigned char *data, bool bWrite)
 {
